requirements.txt
Streamlit
pyrebase4
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2
plotly
python-dotenv